requirements.txt
mss>=9.0.0
pynput>=1.7.6
Pillow>=10.0.0
requests>=2.31.0
pyobjc-framework-Quartz>=10.0
